refactored get_db_uri into database.py
