+++ sds_data_manager/lambda_code/SDSCode/create_schema.py
@@ -2,33 +2,18 @@
 
 import json
 import logging
-import os
 import sys
 
-import boto3
 import requests
 from SDSCode.database.models import Base
 from sqlalchemy import create_engine
+
+from .database.database import get_db_uri
 
 # Logger setup
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
 logging.basicConfig(stream=sys.stdout, level=logging.DEBUG)
-
-
-def get_db_uri():
-    """Create DB URI from secret manager.
-
-    Returns
-    --------
-        str : DB URI
-    """
-    secret_name = os.environ["SECRET_NAME"]
-    session = boto3.session.Session()
-    client = session.client(service_name="secretsmanager")
-    secret_string = client.get_secret_value(SecretId=secret_name)["SecretString"]
-    db_config = json.loads(secret_string)
-    return f'postgresql://{db_config["username"]}:{db_config["password"]}@{db_config["host"]}:{db_config["port"]}/{db_config["dbname"]}'
 
 
 def send_response(event, context, response_status):
