+++ sds_data_manager/lambda_code/SDSCode/indexer.py
@@ -9,6 +9,7 @@ from sqlalchemy import create_engine, inspect
 from sqlalchemy.orm import Session
 
 from .database import models
+from .database.database import get_db_uri
 from .path_helper import FilenameParser
 
 # Logger setup
@@ -17,21 +18,6 @@ logger.setLevel(logging.INFO)
 logging.basicConfig(stream=sys.stdout, level=logging.DEBUG)
 
 s3 = boto3.client("s3")
-
-
-def get_db_uri():
-    """Create DB URI from secret manager.
-
-    Returns
-    --------
-        str : DB URI
-    """
-    secret_name = os.environ["SECRET_NAME"]
-    session = boto3.session.Session()
-    client = session.client(service_name="secretsmanager")
-    secret_string = client.get_secret_value(SecretId=secret_name)["SecretString"]
-    db_config = json.loads(secret_string)
-    return f'postgresql://{db_config["username"]}:{db_config["password"]}@{db_config["host"]}:{db_config["port"]}/{db_config["dbname"]}'
 
 
 def lambda_handler(event, context, db_session=None):
