+++ tests/lambda_endpoints/test_indexer_lambda.py
@@ -3,7 +3,7 @@
 from sds_data_manager.lambda_code.SDSCode import indexer
 
 
-def test_batch_job_event(test_db_uri, db_session):
+def test_batch_job_event(db_session):
     # TODO: replace event with other event source
     # dict. We don't use "Records" anymore. But
     # leaving for now to test database capabilities.
